+++ src/components/Lens/stories/LensEarthSearchNoGeoSearch.story.js
@@ -1,0 +1,31 @@
+import React from 'react';
+
+import Lens from '../../../components/Lens';
+
+import EarthSearchSidebarPanels from './EarthSearchSidebarPanels';
+import {
+  handleEarthSearchUseMapEffect,
+  earthSearchAvailableFilters
+} from './lens-story-util';
+
+const DEFAULT_CENTER = {
+  lat: 0,
+  lng: 0
+};
+
+const LensEarthSearchNoGeoSearch = () => {
+  return (
+    <>
+      <Lens
+        defaultCenter={DEFAULT_CENTER}
+        defaultZoom={2}
+        SidebarComponents={EarthSearchSidebarPanels}
+        useMapEffect={handleEarthSearchUseMapEffect}
+        placeholder="Look stuffs on Earth Data"
+        availableFilters={earthSearchAvailableFilters}
+      />
+    </>
+  );
+};
+
+export default LensEarthSearchNoGeoSearch;