+++ src/hooks/useMap.js
@@ -1,13 +1,16 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useRef } from 'react';
 import L from 'leaflet';
 
 import {
   isValidLeafletElement,
   clearFeatureGroupLayers,
+  clearFeatureGroupLayer,
   currentLeafletRef,
   setMapView,
   centerMapOnGeoJson,
-  addGeoJsonLayer
+  addGeoJsonLayer,
+  addTileLayer,
+  findLayerByName
 } from '../lib/leaflet';
 import { geoJsonFromLatLn, getGeoJsonCenter } from '../lib/map';
 import { isDomAvailable } from '../lib/device';
@@ -34,26 +37,36 @@ const MAP_STATE_DEFAULT = {
 
 const AVAILABLE_MAP_CONTROLS = Object.keys(MAP_CONFIG_DEFAULTS);
 
-let defaultMapFeatureGroup;
+let defaultMapFeaturesGroup;
+let defaultMapOverlaysGroup;
+
 const mapFeatureGroups = [];
 
 if (isDomAvailable()) {
+  defaultMapFeaturesGroup = new L.FeatureGroup();
+  defaultMapOverlaysGroup = new L.FeatureGroup();
+
   mapFeatureGroups.push({
-    id: 'default',
-    featureGroup: new L.FeatureGroup()
+    id: 'default-features',
+    featureGroup: defaultMapFeaturesGroup
   });
-  defaultMapFeatureGroup = mapFeatureGroups[0].featureGroup;
+
+  mapFeatureGroups.push({
+    id: 'default-overlays',
+    featureGroup: defaultMapOverlaysGroup
+  });
 }
 
 export default function useMap (mapSettings = {}) {
-  const { refMap, availableServices = [], projection, draw = {} } = mapSettings;
+  const refMap = useRef();
+
+  const { availableServices = [], projection, draw = {} } = mapSettings;
   const { onCreatedDraw, shapeOptions } = draw;
 
   const defaultMapSettings = buildDefaultMapSettings(mapSettings);
 
   const [mapState, setMapState] = useState(MAP_STATE_DEFAULT);
   const [mapConfig, setMapConfig] = useState(defaultMapSettings);
-  const [mapShape, setMapShape] = useState();
   const [mapServices] = useState(availableServices);
 
   const { defaultCenter, defaultZoom } = mapConfig;
@@ -111,57 +124,6 @@ export default function useMap (mapSettings = {}) {
     };
   }, [refMap]);
 
-  // To make sure we have acccess to the refMap when adding a shape to the map
-  // we allow the app to define a mapShape that will subsequently be added to the
-  // map. If mapShape isn't available or doesn't change, it won't re-run between
-  // renders to avoid incorrect shape layers
-
-  useEffect(() => {
-    if (!mapShape) return;
-
-    const map = currentLeafletRef(refMap);
-
-    if (!isValidLeafletElement(map)) return;
-
-    const {
-      geoJson,
-      shapeOptions: mapShapeOptions = shapeOptions,
-      centerGeoJson,
-      panToShape = true,
-      zoom,
-      clearOtherLayers = true,
-      featureGroup = defaultMapFeatureGroup
-    } = mapShape;
-
-    const geoJsonLayer = addGeoJsonLayer({
-      geoJson,
-      map,
-      featureGroup,
-      options: mapShapeOptions
-    });
-
-    const layersToExclude = [];
-
-    geoJsonLayer.eachLayer(layer => layersToExclude.push(layer));
-
-    if (panToShape) {
-      centerMapOnGeoJson({
-        geoJson: centerGeoJson,
-        map,
-        settings: {
-          zoom
-        }
-      });
-    }
-
-    if (clearOtherLayers) {
-      handleClearLayers({
-        featureGroup,
-        excludeLayers: layersToExclude
-      });
-    }
-  }, [mapShape]);
-
   /**
    * handleOnZoomEnd
    */
@@ -208,7 +170,7 @@ export default function useMap (mapSettings = {}) {
    */
 
   async function handleClearLayers ({
-    featureGroup = defaultMapFeatureGroup,
+    featureGroup = defaultMapFeaturesGroup,
     excludeLayers = []
   } = {}) {
     const map = currentLeafletRef(refMap);
@@ -223,12 +185,25 @@ export default function useMap (mapSettings = {}) {
   }
 
   /**
+   * handleClearLayer
+   */
+
+  function handleClearLayer (settings) {
+    const map = currentLeafletRef(refMap);
+    clearFeatureGroupLayer({
+      map,
+      featureGroup: defaultMapFeaturesGroup,
+      ...settings
+    });
+  }
+
+  /**
    * handleOnLayerCreate
    */
 
   function handleOnLayerCreate ({
     layer,
-    featureGroup = defaultMapFeatureGroup
+    featureGroup = defaultMapFeaturesGroup
   }) {
     const map = currentLeafletRef(refMap);
 
@@ -240,10 +215,14 @@ export default function useMap (mapSettings = {}) {
   }
 
   /**
-   * handleOnLayerCreate
+   * handleAddShapeToMap
    */
 
   function handleAddShapeToMap (settings = {}) {
+    const map = currentLeafletRef(refMap);
+
+    if (!isValidLeafletElement(map)) return;
+
     const errorBase =
       'useMap::handleAddShapeToMap - Failed to add shape to map';
     const {
@@ -253,8 +232,9 @@ export default function useMap (mapSettings = {}) {
       zoom,
       shapeOptions: mapShapeOptions = shapeOptions,
       clearOtherLayers = true,
-      featureGroup
+      featureGroup = defaultMapFeaturesGroup
     } = settings;
+
     let centerGeoJson;
 
     if (!geoJson) {
@@ -267,14 +247,106 @@ export default function useMap (mapSettings = {}) {
       centerGeoJson = getGeoJsonCenter(geoJson);
     }
 
-    setMapShape({
-      shapeOptions: mapShapeOptions,
+    const geoJsonLayer = addGeoJsonLayer({
       geoJson,
-      centerGeoJson,
-      panToShape,
-      zoom,
-      clearOtherLayers,
+      map,
+      featureGroup,
+      options: mapShapeOptions
+    });
+
+    const layersToExclude = [];
+
+    geoJsonLayer.eachLayer(layer => layersToExclude.push(layer));
+
+    if (panToShape) {
+      centerMapOnGeoJson({
+        geoJson: centerGeoJson,
+        map,
+        settings: {
+          zoom
+        }
+      });
+    }
+
+    if (clearOtherLayers) {
+      handleClearLayers({
+        featureGroup,
+        excludeLayers: layersToExclude
+      });
+    }
+  }
+
+  /**
+   * handleClearShapeLayers
+   */
+
+  function handleClearShapeLayers (settings) {
+    handleClearLayers({
+      featureGroup: defaultMapFeaturesGroup,
+      ...settings
+    });
+  }
+
+  /**
+   * handleAddTileLayer
+   */
+
+  function handleAddTileLayerToMap (settings = {}) {
+    const map = currentLeafletRef(refMap);
+
+    if (!isValidLeafletElement(map)) return;
+
+    const {
+      url,
+      options,
+      featureGroup = defaultMapOverlaysGroup,
+      clearOtherLayers
+    } = settings;
+
+    const layer = addTileLayer({
+      url,
+      map,
+      featureGroup,
+      options
+    });
+
+    const layersToExclude = [layer];
+
+    if (clearOtherLayers) {
+      handleClearLayers({
+        featureGroup,
+        excludeLayers: layersToExclude
+      });
+    }
+  }
+
+  /**
+   * handleClearTileLayers
+   */
+
+  function handleClearTileLayers (settings) {
+    handleClearLayers({
+      featureGroup: defaultMapOverlaysGroup,
+      ...settings
+    });
+  }
+
+  /**
+   * handleClearTileLayer
+   */
+
+  function handleClearTileLayer ({
+    featureGroup = defaultMapOverlaysGroup,
+    name
+  } = {}) {
+    const layer = findLayerByName({
+      name,
       featureGroup
+    });
+
+    handleClearLayer({
+      featureGroup,
+      layer: layer
     });
   }
 
@@ -351,7 +423,8 @@ export default function useMap (mapSettings = {}) {
 
   return {
     refMap,
-    mapFeatureGroup: defaultMapFeatureGroup,
+    mapFeatureGroup: defaultMapFeaturesGroup,
+    mapOverlaysGroup: defaultMapOverlaysGroup,
     mapConfig,
     mapState,
     services: mapServices,
@@ -364,6 +437,10 @@ export default function useMap (mapSettings = {}) {
     centerMapOnGeoJson: handleCenterMapOnGeoJson,
     addGeoJsonLayer: handleAddGeoJsonLayer,
     addShapeToMap: handleAddShapeToMap,
+    clearShapeLayers: handleClearShapeLayers,
+    addTileLayerToMap: handleAddTileLayerToMap,
+    clearTileLayers: handleClearTileLayers,
+    clearTileLayer: handleClearTileLayer,
     createFeatureGroup: handleCreateFeatureGroup,
     featureGroupById: handleFeatureGroupById
   };